+++ src/code42cli/cmds/detectionlists/args.py
@@ -1,0 +1,4 @@
+def load_detection_list_args(arg_collection):
+    dl_args = {
+        
+    }