+++ src/code42cli/worker.py
@@ -2,40 +2,15 @@ import traceback
 from threading import Thread, Lock
 
 from code42cli.compat import queue
-
-
-class WorkerGroup(object):
-    def __init__(self, handlers):
-        self._worker_cache = {}
-        self._worker_lock = Lock()
-        self._handlers = handlers
-
-    def add_and_get_worker(self, guid):
-        worker = self._worker_cache.get(guid)
-        if worker:
-            return worker
-
-        with self._worker_lock:
-            worker = self._worker_cache.get(guid)
-            if worker:
-                return worker
-            new_worker = Worker(4, self._handlers)
-            self._worker_cache.update({guid: new_worker})
-            return new_worker
-
-    def wait_all(self):
-        for guid in self._worker_cache:
-            worker = self._worker_cache[guid]
-            worker.wait()
+from code42cli.util import print_error
 
 
 class Worker(object):
-    def __init__(self, thread_count, handlers):
+    def __init__(self, thread_count):
         self._queue = queue.Queue()
         self._thread_count = thread_count
         self.__started = False
         self.__start_lock = Lock()
-        self._handlers = handlers
 
     def do_async(self, func, *args, **kwargs):
         if not self.__started:
@@ -43,23 +18,22 @@ class Worker(object):
                 if not self.__started:
                     self.__start()
                     self.__started = True
-        self._queue.put({"func": func, "args": args, "kwargs": kwargs})
+        self._queue.put({u"func": func, u"args": args, u"kwargs": kwargs})
 
     def wait(self):
         self._queue.join()
-        self.__started = False
 
     def _process_queue(self):
         while True:
             try:
                 task = self._queue.get()
-                func = task["func"]
-                args = task["args"]
-                kwargs = task["kwargs"]
+                func = task[u"func"]
+                args = task[u"args"]
+                kwargs = task[u"kwargs"]
                 func(*args, **kwargs)
             except Exception:
                 trace = traceback.format_exc()
-                self._handlers.handle_error(trace)
+                print_error(u"Failed to process row. Trace: {0}".format(trace))
             finally:
                 self._queue.task_done()