+++ CHANGELOG.md
@@ -13,7 +13,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 ### Added
 
 - `code42 profile create` command.
-- `code42 detectionlists departingemployee commands`:
+- `code42 detectionlists departingemployee` commands:
     - `init-csv`
     - `add`